Silence extractor output
